Added parser. Reorganized files.
