+++ scripts/bfasst/compare_waveforms/analyze_graph.py
